session_backend key added to config
