+++ csm/conf/etc/csm/csm.conf
@@ -128,6 +128,7 @@ CSM:
     username: "csm"
     password: ""
     password_decryption_key: "cortx"
+    session_backend: "in_memory"
 
 UDS:
     url: "https://127.0.0.1:5000"
