+++ csm/core/agent/csm_agent.py
@@ -106,6 +106,7 @@ class CsmAgent:
         auth_service = AuthService()
         user_manager = UserManager(db)
         role_manager = RoleManager(roles)
+        Conf.get(const.CSM_GLOBAL_INDEX, const.SESSION_BACKEND_KEY)
         session_manager = SessionManager(const.IN_MEMORY, db)
         CsmRestApi._app.login_service = LoginService(auth_service,
                                                      user_manager,
