+++ csm/core/blogic/const.py
@@ -947,5 +947,6 @@ SETUP_ERROR             = 'SetupError'
 HTTP_ERROR              = 'HttpError'
 
 # Session Backend factory
-DB        = 'database'
-IN_MEMORY = 'in_memory'
+DB                  = 'database'
+IN_MEMORY           = 'in_memory'
+SESSION_BACKEND_KEY = 'CSM>session_backend'
